requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0  
pymongo==4.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
pandas==2.2.2        
matplotlib==3.10.3
seaborn==0.13.2